Replace uint with int for magnitudes
Although uint prevents us from manually coding negative values that
should be positive (such as lengths), it does not prevent underflow when
subtracting from 0. ints are therefore preferable, since it is clearer
to check for values < 0, rather than values of MaxUint.

+++ cmd/main.go
@@ -53,8 +53,8 @@ func run() (err error) {
 		info.Main.Version,
 		terminalWriter,
 		renderer.Screen{
-			Width:  uint(w),
-			Height: uint(h),
+			Width:  w,
+			Height: h,
 		},
 	)
 
@@ -69,8 +69,8 @@ func run() (err error) {
 		keyReader,
 		renderer,
 		editor.Config{
-			Width:  uint(w),
-			Height: uint(h),
+			Width:  w,
+			Height: h,
 		},
 		logger,
 	)
